backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pyyaml==6.0.1
requests==2.31.0
pytz==2023.3
APScheduler==3.10.4
celery==5.3.4
redis==5.0.1
email-validator==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
cors==1.0.1
flask-cors==4.0.0
